Moved styled components to proper folder
